feat: 공통으로 적용될 page title component 구현
