--- src/app/(detail)/jobs/page.tsx
@@ -1,7 +1,0 @@
-function JobsPage() {
-  return (
-    <div>JobsPage</div>
-  );
-}
-
-export default JobsPage;
