--- src/app/(detail)/organizers/page.tsx
@@ -1,7 +1,0 @@
-function OrganizersPage() {
-  return (
-    <div>OrganizersPage</div>
-  );
-}
-
-export default OrganizersPage;
