--- src/app/(detail)/projects/page.tsx
@@ -1,7 +1,0 @@
-function ProjectsPage() {
-  return (
-    <div>ProjectsPage</div>
-  );
-}
-
-export default ProjectsPage;
