--- src/app/(detail)/reviews/page.tsx
@@ -1,7 +1,0 @@
-function ReviewsPage() {
-  return (
-    <div>ReviewsPage</div>
-  );
-}
-
-export default ReviewsPage;
