--- src/app/blog/page.tsx
@@ -1,7 +1,0 @@
-function BlogPage() {
-  return (
-    <div>BlogPage</div>
-  );
-}
-
-export default BlogPage;
