+++ src/app/jobs/page.tsx
@@ -1,0 +1,13 @@
+import PageTitle from '@/components/molecules/PageTitle';
+import ShareAlarmSection from '@/components/molecules/ShareAlarmSection';
+
+function JobsPage() {
+  return (
+    <>
+      <PageTitle title="채용" subTitle="지금, 열정이 넘치는 참가자가 가득한 DND의 채용 공고를 확인하세요!" />
+      <ShareAlarmSection />
+    </>
+  );
+}
+
+export default JobsPage;
