+++ src/app/organizers/page.tsx
@@ -1,0 +1,13 @@
+import PageTitle from '@/components/molecules/PageTitle';
+import ShareAlarmSection from '@/components/molecules/ShareAlarmSection';
+
+function OrganizersPage() {
+  return (
+    <>
+      <PageTitle title="운영진" subTitle="DND활동을 실현하게 한 멤버들을 소개합니다." />
+      <ShareAlarmSection />
+    </>
+  );
+}
+
+export default OrganizersPage;
