+++ src/app/projects/[id]/page.tsx
