+++ src/app/projects/page.tsx
@@ -1,0 +1,13 @@
+import PageTitle from '@/components/molecules/PageTitle';
+import ShareAlarmSection from '@/components/molecules/ShareAlarmSection';
+
+function ProjectsPage() {
+  return (
+    <>
+      <PageTitle title="프로젝트" subTitle="참가자들의 다양한 프로젝트를 확인해 보세요!" />
+      <ShareAlarmSection />
+    </>
+  );
+}
+
+export default ProjectsPage;
