+++ src/app/reviews/page.tsx
@@ -1,0 +1,13 @@
+import PageTitle from '@/components/molecules/PageTitle';
+import ShareAlarmSection from '@/components/molecules/ShareAlarmSection';
+
+function ReviewsPage() {
+  return (
+    <>
+      <PageTitle title="후기" subTitle="지금 참가자들의 따끈한 후기를 확인해 보세요!" />
+      <ShareAlarmSection />
+    </>
+  );
+}
+
+export default ReviewsPage;
