+++ src/components/global/TopNavigationMenu/index.module.scss
@@ -87,3 +87,14 @@
     }
   }
 }
+
+.blogLink {
+  display: flex;
+  flex-direction: row;
+  gap: 4px;
+  align-items: center;
+
+  &:hover {
+    text-decoration: none;
+  }
+}
